Update render to allow string or null vnode
This is needed for recursive rendering. It also gives us a happy path to
using this API when we don't have all the properties available. Like
Lit, null or undefined will result in an empty string node. String will
result in a string node for the value of the string. A vnode will be
constructed using the internal function renderVNode.