Add known-tags registry
...to break circular module dependency. Previously we were defining the
set of blessed tags using static imports and exports. Render would then
use this static list. However, some of the tags in the blessed list
themselves used render to define themselves. So we had a circular
dependency problem. THe known-tags registry resolves this by allowing
for the registration of tags after definition.