Generate a cid for repeat elements if no id
IDs enable efficient re-rendering of dynamic lists. However, if a
model does not provide an ID, we should just generate one for the
purpose of making the logic work. Those items will not get efficient
re-rendering (they will be destroyed and re-created), but it will still
work. This is the same tradeoff React makes.