Allow mustache-style views, or html tag fn
- Renderable renamed to View.
- Node renamed to VNode (in keeping with common-ui).

A view is a combination of a template (vnode) and context (key-value
map of reactive values). You can create them with the
`view(markup, context)` function, which takes a markup string with
mustaches, and a context object, returning a view.

Tagged template literal `html` is a shortcut for creating views. It will
automatically generate keys for each value in a tagged template literal
hole, so you don't need to name them.