chore: Add Dockerfile and compose.yml for planning-server
This commit adds a Dockerfile and a compose.yml file for the planning-server package. The Dockerfile sets up the environment using the specified Deno version and exposes the specified port. The compose.yml file defines the services and their configurations for the planning-server container. These additions will facilitate the deployment and containerization of the planning-server package.