Work around broken container queries vs slots
Introduce ResponsiveElement helper class which uses a resize observer to
generate breakpoint information about the container context. This is to
work around the fact that container queries are currently broken for
slotted content. The spec is updated and browsers are fixing it, but
it's going to be a bit before there is broad support for slotted
container queries.

See https://chromestatus.com/feature/5242724333387776