hack to remove a particular foot gun around streams (#1297)
When the order of unnamed cells changes, a stream might be assigned to a previously existing value, typically what was the output for a derive/lift. That means `$stream: true` won't be set, and so it won't be treated as stream. This fixes this particular case by namespacing the streams away. Still not great overall.