Client-side schema queries to limit docs fetched. (#1269)
Remove crawler code, so we don't need to fetch reachable, but not schema-reachable docs.
Use local queries to determine docs needed for a cell/schema
Provide schema when syncing charm cells in CharmManager. Minor type cleanup.
Load source links when we load docs on the server in loadFactsForDoc, and on the client in querySchema. Also do this in traverse when we followPointer. The former two capture the initial doc's source, while the followPointer version handles any that are referenced.
Load recipe links in the same places
Remove _batchForDoc, _processCurrentBatch, _addToBatch, _runBatchesUntilSettled, _ensureIsSyncedf, and _ensureIsSyncedById. Also changed doc loading to account for schema, so we can load the same doc with different schema, but a second request with the same schema will wait on the pending request.
Added toAddress method to BaseObjectManager. Implementations should use this to convert a string to whatever form of address they use. Also added getTarget method that gets an address object from a URI.
Added an entityIdStr method to doc-map.ts, since getting the string version is tricky, and we do it in many places. This is exported from runner.
Changed some classes used to abstract the details of interacting with how objects are stored in query.ts
Made ClientObjectManager abstract. Expose the readValues and missingDocs contents, as well.
Added StoreObjectManager, which is backed by the storage systems heap. When loading a doc, add that doc to the set of readValues, so we track the usage (this should have been done before, but wasn't).
Added DocObjectManager, which is backed by the DocumentMap and the StorageProvider, so we can run schema queries against what we have locally.
Remove the readDependentDocs/writeDependentDocs maps from the ObjectManager, since we aren't using this, and aren't properly maintaining its contents
Added tracking of server subscriptions, so we don't need to re-issue any current subscriptions. This can also be used if we reconnect, so that we can re-establish our list of subscriptions (Ellyse implemented the reconnect logic).
Added tracking of pending nursery changes, and be less aggressive about purging the nursery. When we receive updates from the server, we only send them to the subscribers if the update doesn't correspond to a pending update in the nursery (otherwise, we already have more recent data in the doc). This code can go away when we have our transaction system integrated.
Switched some traversal code to use the normalized links. Use the new isAnyCellLink and parseLink methods instead of isPointer and getPointerInfo.
Moved MapSet from util to traverse, and added some more functionality.