Change StorageProvider.sync to take a SchemaPathSelector. (#1513)
* Change StorageProvider.sync to take a SchemaPathSelector.
Removed the expectedInStorage and schemaContext arguments to StorageProvider.sync, since the former was unused and the latter is replaced by selector.
Changed the structures that tracked the SchemaContext in cache to use SchemaPathSelectors instead.
Renamed Replia.getSchema to generateSchemaFromLabels to better indicate the function's purpose.
Updated curl to new api that takes URI instead of EntityId.

* Added integration test