Fix various issues causing arrays to have conflicts or sometimes be emptied (#1644)
* Filter out redundant fact updates before sending them to the server.
Be more consistent with cleaning up the nursery before we update the heap, so that any heap subscribers won't see stale values in the nursery.

* Expand the changedFacts filtering, since the facts we get from the transaction will have invalid cause chains (if we're not writing redundant facts, their cause chain won't progress).

* Added comments to Chronicle to explain things better
Changed Chronicle.commit to generate claims instead of assertions when the value has not changed.
Changed Novelty.claim to iterate over a copy of the candidates, since we modify the collection in the loop. Reversed the Address.includes parameters in the test in that loop, since we are trying to remove child candidates, and not parent candidates.

* Added comment about retract unclaimed not really being a valid test.

* When running the map function, instead of just setting the child elements, also set the full array value. (#1643)

* When running the map function, instead of just setting the child elements, also set the full array value.
When setting an array that is shorter, include a change to the array's length that shortens it.

* Use get instead of getRaw
Use splice instead of setting length