Fix memory route for subscribe to work on firefox (#1697)
* Fix memory route for subscribe to work on firefox
Firefox sends `Connection: keep-alive, Upgrade` as the header, but the route was only matching `Upgrade` (which is what Chrome sends). This header is a list of options, so use a regex to check that it's one of the options in the list instead.

* Make the check of the Connection header case insensitive per spec.
Also since the protocol allows more complex values for the Upgrade header, handle the options to have a list, an optioanal protocol_version, and be case-insensitive (recipient SHOULD).

---------

Co-authored-by: Jordan Santell <jsantell@users.noreply.github.com>