chore: Specify API host for runtime, deriving other addresses as needed (#1818)
chore: Specify API host for runtime, deriving other addresses as needed, and check in shell that runtime is live.

Currently, when the runtime connects to an incorrect/invalid server
(e.g. forgetting to run local server), the web socket connection retries
until success, leaving the runtime initialization hanging. A general
"API host" URL is needed for the runtime in order to ping the status,
which would circumvent this socket loop scenario and alert that the
server is invalid.

With an API host URL, other runtime values can be derived, removing
redundant params.

With static asset cache no longer explicit, and all but runtime usage of
StaticCache being in a Deno environment, the StaticCache functionality
now explicitly uses HTTP or the filesystem (StaticCacheHTTP and
StaticCacheFS).